Get rid of some debug stuff in Input that was causing debug output spew and slowdowns
Transform returned positions in the EditorMesh position return stuff

Update indoormesh delta times to take a float, have all that timey stuff worked out now

Ditto Particle delta stuff

Dot extension method for half4 to vec3 dots

All deltatimes are now floats and all movers now work from seconds

New UpdateTimer class that handles all the goblinry of fixed time steps, remainders, and variable time steps in a simple class

All test programs updated to work with UpdateTimer

+++ MeshLib/IndoorMesh.cs
@@ -352,9 +352,9 @@ namespace MeshLib
 		}
 
 
-		public void Update(int msDelta)
+		public void Update(float msDelta)
 		{
-			Debug.Assert(msDelta > 0);
+			Debug.Assert(msDelta > 0f);
 
 			UpdateAnimatedLightMaps(msDelta);
 		}
@@ -643,7 +643,7 @@ namespace MeshLib
 		}
 
 
-		void UpdateAnimatedLightMaps(int msDelta)
+		void UpdateAnimatedLightMaps(float msDelta)
 		{
 			List<string>	mats	=mMatLib.GetMaterialNames();
